BL: Adapt to Victors changes in Makefile + add regression tests for OMM (not all are working yet)

+++ ELSI/regression/regression.sh
@@ -10,6 +10,14 @@ else
    echo "construct_and_write_elpa successfull"
    rm elsi_eigenvalue_problem.hdf5
 fi
+mpirun -n 4 ../../bin/construct_and_write_omm 100 10 > /dev/null
+fail=`h5diff elsi_eigenvalue_problem.hdf5 data/elsi_eigenvalue_problem.hdf5`
+if [ "x$fail" != "x" ]; then
+   echo "construct_and_write_omm failed"
+else
+   echo "construct_and_write_omm successfull"
+   rm elsi_eigenvalue_problem.hdf5
+fi
 cd ..
 
 # Set and Write
@@ -20,6 +28,14 @@ if [ "x$fail" != "x" ]; then
    echo "set_and_write_elpa failed"
 else
    echo "set_and_write_elpa successfull"
+   rm elsi_eigenvalue_problem.hdf5
+fi
+mpirun -n 4 ../../bin/set_and_write_omm 100 10 > /dev/null
+fail=`h5diff elsi_eigenvalue_problem.hdf5 data/elsi_eigenvalue_problem.hdf5`
+if [ "x$fail" != "x" ]; then
+   echo "set_and_write_omm failed"
+else
+   echo "set_and_write_omm successfull"
    rm elsi_eigenvalue_problem.hdf5
 fi
 cd ..
@@ -34,6 +50,14 @@ else
    echo "read_and_write_elpa successfull"
    rm elsi_eigenvalue_problem_out.hdf5
 fi
+mpirun -n 4 ../../bin/read_and_write_omm > /dev/null
+fail=`h5diff elsi_eigenvalue_problem_out.hdf5 data/elsi_eigenvalue_problem_out.hdf5`
+if [ "x$fail" != "x" ]; then
+   echo "read_and_write_omm failed"
+else
+   echo "read_and_write_omm successfull"
+   rm elsi_eigenvalue_problem_out.hdf5
+fi
 cd ..
 
 # Read and Solve
@@ -46,6 +70,16 @@ if [ "x$fail" != "x" ]; then
    echo "read_and_solve_elpa failed"
 else
    echo "read_and_solve_elpa successfull"
+   rm eigenvalues.dat
+fi
+mpirun -n 4 ../../bin/read_and_solve_omm 100 > buffer.dat
+tail -n 100 buffer.dat > eigenvalues.dat
+rm buffer.dat
+fail=`diff eigenvalues.dat data/eigenvalues.dat`
+if [ "x$fail" != "x" ]; then
+   echo "read_and_solve_omm failed"
+else
+   echo "read_and_solve_omm successfull"
    rm eigenvalues.dat
 fi
 cd ..
